fix: Rename all refs to Polaris into Adevinta in CatalogApp
